pyproject.toml
[project]
name = "fractaltrainer"
version = "0.1.0"
description = "Fractal-dimension-guided hyperparameter repair loop"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Vegar", email = "vegarjr@users.noreply.github.com"},
]
keywords = [
    "fractal-dimension",
    "continual-learning",
    "hyperparameter-tuning",
    "self-modifying-ai",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "anthropic>=0.40.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]
notebook = [
    "matplotlib>=3.7.0",
    "jupyter>=1.0.0",
]

[project.scripts]
fractaltrainer = "fractaltrainer.cli:main"

[project.urls]
Homepage = "https://github.com/vegarjr/FractalTrainer"
Repository = "https://github.com/vegarjr/FractalTrainer"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"
